pyproject.toml
[project]
name = "citrasense"
version = "1.0.0"
description = "Remotely control a telescope while it polls for tasks, collects and edge processes data, and delivers results and data for further processing."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{ name = "Patrick McDavid", email = "patrick@citra.space" }]
keywords = [
  "astronomy",
  "telescope",
  "astrophotography",
  "INDI",
  "NINA",
  "KStars",
  "observatory",
  "imaging",
  "remote-telescope",
  "telescope-control",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Scientific/Engineering :: Astronomy",
  "Topic :: System :: Hardware :: Hardware Drivers",
  "Operating System :: OS Independent",
]
dependencies = [
  "click",
  "requests",
  "httpx",
  "python-dateutil",
  "skyfield",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "websockets>=13.0",
  "platformdirs>=4.0.0",
  "jinja2>=3.1.0",
  "ntplib>=0.4.0",
  "numpy>=2.0.0",
  "astropy>=6.0.0",
  "pandas>=2.0.0",
  "scipy>=1.14.0",
  "sep>=1.3.0",
  "keplemon>=3.6.3",
  "astropy-healpix>=1.0.0",
  "astroquery>=0.4.6",
  "Pillow>=10.0.0",
]

[project.scripts]
citrasense = "citrasense.__main__:cli"

[project.urls]
Homepage = "https://citra.space"
Documentation = "https://docs.citra.space/citrascope/"
Repository = "https://github.com/citra-space/citrascope"
Issues = "https://github.com/citra-space/citrascope/issues"
Changelog = "https://github.com/citra-space/citrascope/releases"

[project.optional-dependencies]
indi = [
  "pyindi-client",
]
usb-camera = [
  "opencv-python>=4.8.0",
  "cv2-enumerate-cameras>=1.0.0",
]
rpi = [
  "picamera2>=0.3.12; sys_platform == 'linux'",
]
ximea = []
zwo-mount = [
  "pyserial>=3.5",
]
all-hardware = [
  "citrasense[indi,zwo-mount]",
  "opencv-python>=4.8.0",
  "cv2-enumerate-cameras>=1.0.0",
  "picamera2>=0.3.12; sys_platform == 'linux'",
]
kstars = [
  "dbus-python",
]
all = [
  "citrasense[indi,kstars]",
]
dev = [
  "black",
  "pre-commit",
  "ruff",
  "pyright",
  "types-requests",
  "bump-my-version",
  "citrasense[test]",
]
test = ["pytest", "pytest-cov", "mockito"]
build = ["build"]
deploy = ["twine"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# python-prctl is Linux-only and can't be built on macOS/Windows during resolution.
dependency-metadata = [
  { name = "python-prctl", version = "1.8.1", requires-python = ">=3.0" },
]

[tool.black]
line-length = 120
force-exclude = '''
^/(
  (
      \.eggs
    | \.git
    | \.pytest_cache
    | \.tox
    | dist
    | data
  )/
)
'''

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "PT", "B", "T20", "C4", "RUF"]
ignore = ["E203", "RUF001", "RUF002", "RUF003"]

[tool.ruff.lint.isort]
known-first-party = ["citrasense"]

[tool.ruff.lint.per-file-ignores]
"citrasense/settings/citrasense_settings.py" = ["E402"]
"tests/unit/test_msi_processors.py" = ["E402"]
"tests/*" = ["RUF012"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = [
  "--cov=citrasense",
  "--cov-report=term-missing",
  "--cov-report=xml",
  "--cov-branch",
  "--import-mode=importlib",
  "-m", "not slow",
]
markers = [
  "slow: expensive tests (plate solving, real FITS processing, etc.) — skipped locally, run in CI",
  "integration: requires live hardware or services",
]

[tool.coverage.run]
omit = [
    "tests/*",
    "__init__.py",
    "citrasense/hardware/devices/camera/moravian_camera.py",
    "citrasense/hardware/devices/camera/ximea_camera.py",
    "citrasense/hardware/devices/camera/rpi_hq_camera.py",
    "citrasense/hardware/devices/camera/usb_camera.py",
    "citrasense/hardware/devices/focuser/zwo_eaf.py",
    "citrasense/hardware/devices/mount/zwo_am_mount.py",
    "citrasense/hardware/devices/mount/zwo_am_transport.py",
    "citrasense/hardware/indi/*",
    "citrasense/hardware/kstars/*",
    "citrasense/hardware/nina/*",
    "citrasense/location/gps_monitor.py",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.bumpversion]
current_version = "1.0.0"
commit = true
tag = true

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'
